Folder Path corrections OMS.ini issue fix Masters - Button correction Combo box style change Excel corrections Description fix in Indent Master Screen

+++ LoginWindow.xaml.cs
@@ -86,9 +86,11 @@ namespace OrderManagementTool
                 string password = txt_password.Password.ToString().Split(new string[] { "\r\n" }, StringSplitOptions.None)[0];
                 DirectoryInfo dInfo = new DirectoryInfo(path);
                 FileInfo[] files = null;
+                bool isFileExist = false;
                 try
                 {
                     files = dInfo.GetFiles("OMT.ini");
+                    isFileExist = true;
                 }
                 catch (Exception exc)
                 {
@@ -99,8 +101,11 @@ namespace OrderManagementTool
                 }
                 try
                 {
-                    string[] lines = { username + "," + password };
-                    File.WriteAllLines(path + "\\OMT.ini", lines);
+                    if (!isFileExist)
+                    {
+                        string[] lines = { username + "," + password };
+                        File.WriteAllLines(path + "\\OMT.ini", lines);
+                    }
                 }
                 catch (Exception ex)
                 {
